==> 02-Prepare_Categories.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
# Ignore warnings
warnings.filterwarnings("ignore")

# lines = False ( default value )
#          assumption is that the entire content of the JSON file represents a single JSON object.
# lines = True
#         each line in the JSON file is treated as a separate JSON object.
data = pd.read_json('Cleaned_Data.json', lines=True)

categories_to_keep = ['POLITICS', 'WELLNESS', 'ENTERTAINMENT', 
                      'TRAVEL', 'STYLE & BEAUTY' , 'QUEER VOICES' ,
                      'FOOD & DRINK' ,'BUSINESS' , 'SPORTS']

Categories_Map = {}

for ele in data['category'].unique():
    if ele in categories_to_keep:
        Categories_Map.update({ele:ele})

data['category'].nunique()

def Update_map(li,label):
    for ele in li:
        Categories_Map.update({ele:label})

#################################################################################################

Update_map(['HEALTHY LIVING','IMPACT', 'WOMEN' , 'FIFTY' ] , 'OTHER1')

Update_map([ 'ARTS & CULTURE' , 'ARTS' , 'CULTURE & ARTS', 'TECH' , 'LATINO VOICES' , 'MEDIA' , 'COMEDY' ] , 'OTHER2')

Update_map(['BLACK VOICES' , 'U.S. NEWS' , 'CRIME' , 'WEIRD NEWS' , 'GOOD NEWS' ] , 'OTHER3')

Update_map(['DIVORCE' , 'WEDDINGS' , 'ENVIRONMENT'  , 'HOME & LIVING' ] , 'OTHER4')

Update_map(['MONEY' , 'EDUCATION' , 'COLLEGE' , 'STYLE' , 'TASTE' , 'SCIENCE' , 'GREEN'] , 'OTHER5')

Update_map(['PARENTING' , 'PARENTS'] , 'OTHER6')

Update_map(['RELIGION' , 'WORLDPOST' , 'WORLD NEWS' ,'THE WORLDPOST'] , 'OTHER7')

#################################################################################################

def Re_categorize(cat):
    return Categories_Map[cat]

data.loc[:, ('category')] = data.loc[:,('category')].apply(lambda x: Re_categorize(x))

data[data['category'].isin(categories_to_keep)]['category'].value_counts()

data[~data['category'].isin(categories_to_keep)]['category'].value_counts()

data.to_json('Cleaned_Data.json', orient='records', lines=True)
