==> 03-Split_Data.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.model_selection import train_test_split
import warnings
# Ignore warnings
warnings.filterwarnings("ignore")

# lines = False ( default value )
#          assumption is that the entire content of the JSON file represents a single JSON object.
# lines = True
#         each line in the JSON file is treated as a separate JSON object.
data = pd.read_json('Cleaned_Data.json', lines=True)

X_train_plus_valid, X_test, y_train_plus_valid, y_test= train_test_split(
    data['headline'], data['category'], random_state=42, test_size = 0.2, stratify=data['category'])


print(X_train_plus_valid.shape)
print(y_train_plus_valid.shape)
print(X_test.shape)
print(y_test.shape)

X_train, X_val, y_train, y_val = train_test_split(
    X_train_plus_valid, y_train_plus_valid, random_state=42, test_size = 0.25,
    stratify=y_train_plus_valid)

print(X_train.shape)
print(y_train.shape)
print(X_val.shape)
print(y_val.shape)

categories_to_keep = ['POLITICS', 'WELLNESS', 'ENTERTAINMENT', 
                      'TRAVEL', 'STYLE & BEAUTY' , 'QUEER VOICES' ,
                      'FOOD & DRINK' ,'BUSINESS' , 'SPORTS']

def process_test_val(li):
    templi = []
    for label in li:
        if label in categories_to_keep:
            templi.append(label)
        else:
            templi.append('OTHER')
    return pd.Series(templi)

y_test = process_test_val(y_test)

y_val = process_test_val(y_val)

# Convert the list of strings to a DataFrame
df = pd.DataFrame(list(X_train), columns=['Data'])

# Specify the CSV file path
csv_file_path = "Spliting Data\\X_train.csv"

# Write DataFrame to CSV file
df.to_csv(csv_file_path, index=False)

# Convert the list of strings to a DataFrame
df = pd.DataFrame(list(y_train), columns=['Data'])

# Specify the CSV file path
csv_file_path = "Spliting Data\\y_train.csv"

# Write DataFrame to CSV file
df.to_csv(csv_file_path, index=False)

# Convert the list of strings to a DataFrame
df = pd.DataFrame(list(X_test), columns=['Data'])

# Specify the CSV file path
csv_file_path = "Spliting Data\\X_test.csv"

# Write DataFrame to CSV file
df.to_csv(csv_file_path, index=False)

# Convert the list of strings to a DataFrame
df = pd.DataFrame(list(y_test), columns=['Data'])

# Specify the CSV file path
csv_file_path = "Spliting Data\\y_test.csv"

# Write DataFrame to CSV file
df.to_csv(csv_file_path, index=False)

# Convert the list of strings to a DataFrame
df = pd.DataFrame(list(X_val), columns=['Data'])

# Specify the CSV file path
csv_file_path = "Spliting Data\\X_val.csv"

# Write DataFrame to CSV file
df.to_csv(csv_file_path, index=False)

# Convert the list of strings to a DataFrame
df = pd.DataFrame(list(y_val), columns=['Data'])

# Specify the CSV file path
csv_file_path = "Spliting Data\\y_val.csv"

# Write DataFrame to CSV file
df.to_csv(csv_file_path, index=False)
