==> 04-Balance_Data.ipynb <==
import csv
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.cluster import KMeans
from sklearn.feature_extraction.text import CountVectorizer
import numpy as np
import nlpaug.augmenter.word as naw
import warnings
# Ignore warnings
warnings.filterwarnings("ignore")

X_train = pd.read_csv('Spliting Data\\X_train.csv')
y_train = pd.read_csv('Spliting Data\\y_train.csv')

X_train = X_train['Data']
y_train = y_train['Data']

train_dataframe = pd.DataFrame ({'headline': X_train , 'category' :y_train })

train_dataframe.head()

def take_most_important_samples(df, percent , n_clusters ):
    df_filtered = pd.DataFrame ()

    for i in range (n_clusters):
        temp = df[df['cluster_label'] == i]
        sz = round(percent * temp.shape[0] )
        if sz :
            df_filtered = pd.concat([df_filtered, temp.head(sz)], ignore_index=True)
    return df_filtered

def reduce(category , n_clusters , percent ):
    
    global train_dataframe
    X = train_dataframe[train_dataframe['category'] == category] 
    X = X['headline']
    
    print ("category shape : " ,X.shape) 
    
    vectorizer = CountVectorizer()
    sparse_matrix = vectorizer.fit_transform(X)
    print ("sparse_matrix shape : ",sparse_matrix.shape) 
    
    kmeans = KMeans(n_clusters=n_clusters, random_state=42)
    cluster_labels = kmeans.fit_predict(sparse_matrix)
    print ( "cluster_labels shape : " , cluster_labels.shape)
    
    X = pd.DataFrame ({'headline': X , 'cluster_label' :cluster_labels })
    centroids = kmeans.cluster_centers_
    distances_to_centroids = np.linalg.norm(sparse_matrix - centroids[cluster_labels], axis=1)
    importance_values =  1 / (1 + distances_to_centroids )
    print ("importance_values shape : " , importance_values.shape)
    
    X = pd.DataFrame ({'headline': X['headline'] , 'cluster_label' :cluster_labels , "importance" : importance_values })
    print("category shape : ", X.shape)
    
    X = X.groupby('cluster_label').apply(lambda group: group.sort_values(by='importance', ascending=False)).reset_index(drop=True)
    X_filtered = take_most_important_samples(X, percent, n_clusters )
    
    train_dataframe =  train_dataframe[train_dataframe['category'] != category ]
    X_filtered = pd.DataFrame( {'headline': X_filtered['headline'] ,"category": [category] * X_filtered.shape[0] })
    train_dataframe = pd.concat([train_dataframe,X_filtered ], ignore_index=True)
    
    print ("New train_dataframe shape : " , train_dataframe.shape )
    print ("--------------------------------------------------\n\n")

reduce ('POLITICS' , 80 , 0.8)

train_dataframe['category'].value_counts()

aug = naw.SynonymAug()

def augment0_Txt(text):
    return aug.augment(text)
    
def augmentation ( category , target):
    flag = False
    global train_dataframe 
    data_temp = train_dataframe[train_dataframe['category'] == category]
        
    print("category: ", category)
    print("Number of headlines : " , data_temp.shape[0])

    ndata_temp = data_temp.copy()
    if ndata_temp.shape[0] >= target:
        return flag
    for headline in data_temp['headline']:
        if ndata_temp.shape[0] >= target:
            break
        txt1 = augment0_Txt(headline)
        if txt1:
            flag = True
            txt1= txt1[0]
            new_row1 = {'headline': txt1, 'category': category}
            ndata_temp = pd.concat([ndata_temp, pd.DataFrame([new_row1])], ignore_index=True)

    train_dataframe = train_dataframe[train_dataframe['category'] != category]
    print("Number of headlines after augmentation : " , ndata_temp.shape) 
    print ("-----------------------------------------------------------")
    train_dataframe = pd.concat([train_dataframe,ndata_temp])
    return flag

while True:
    ans = False
    for cat in train_dataframe['category'].unique():
        ans |= augmentation(cat,16811)
        
    print("Number of duplicate headlines = ", train_dataframe.duplicated(['headline'], keep = 'last').sum())

    train_dataframe.drop_duplicates(['headline'], keep = 'last', inplace = True)
    
    print("Total number of rows after removing duplicates = ",train_dataframe.shape[0])
    
    if ans == False:
        break

train_dataframe['category'].value_counts()

X_train = train_dataframe['headline']
y_train = train_dataframe['category']

# Convert the list of strings to a DataFrame
df = pd.DataFrame(list(X_train), columns=['Data'])

# Specify the CSV file path
csv_file_path = "Spliting Data\\X_train.csv"

# Write DataFrame to CSV file
df.to_csv(csv_file_path, index=False)

# Convert the list of strings to a DataFrame
df = pd.DataFrame(list(y_train), columns=['Data'])

# Specify the CSV file path
csv_file_path = "Spliting Data\\y_train.csv"

# Write DataFrame to CSV file
df.to_csv(csv_file_path, index=False)
