==> 05-preprocessing_NER.ipynb <==
import pandas as pd
import nltk
from nltk.stem import WordNetLemmatizer
from nltk import word_tokenize, pos_tag, ne_chunk
from nltk.corpus import stopwords

import warnings
# Ignore warnings
warnings.filterwarnings("ignore")

X_train = pd.read_csv ('..\\Spliting Data\\X_train.csv')['Data']
y_train = pd.read_csv ('..\\Spliting Data\\y_train.csv')['Data']

X_test = pd.read_csv  ('..\\Spliting Data\\X_test.csv')['Data']
y_test = pd.read_csv ('..\\Spliting Data\\y_test.csv')['Data']

X_val = pd.read_csv   ('..\\Spliting Data\\X_val.csv')['Data']
y_val = pd.read_csv ('..\\Spliting Data\\y_val.csv')['Data']

print(X_train.shape)
print(y_train.shape)
print('-----------------')
print(X_test.shape)
print(y_test.shape)
print('-----------------')
print(X_val.shape)
print(y_val.shape)


stop_words = set(stopwords.words('english'))

def preprocess(text):
    
    tokens = nltk.word_tokenize(text)  # Tokenize the text into word tokens
    
    tokens = [token for token in tokens if token.isalpha()]  # Remove punctuation and numbers
    
    POS = pos_tag(tokens)
    
    NER_tree = ne_chunk(POS)
    
    ner = [] 
    
    for entity in NER_tree:
        if isinstance(entity, nltk.Tree):
            word = entity.leaves()[0][0].lower()
            if word not in stop_words:
                ner.append(word)
    
    
    return ' '.join(ner)

# X_train = [preprocess(text) for text in X_train]

X_test = [preprocess(text) for text in X_test]

X_val = [preprocess(text) for text in X_val]

# train_dataframe = pd.DataFrame ({'headline': X_train , 'category' :y_train })

val_dataframe = pd.DataFrame ({'headline': X_val , 'category' :y_val })

test_dataframe = pd.DataFrame ({'headline': X_test , 'category' :y_test })

# print("Number of duplicate headlines in train = ", train_dataframe.duplicated(['headline'], keep = 'last').sum())
# print("Number of duplicate headlines in val   = ", test_dataframe.duplicated(['headline'], keep = 'last').sum())
# print("Number of duplicate headlines in val   = ", val_dataframe.duplicated(['headline'], keep = 'last').sum())

#Removing duplicates
#inplace=True: This argument modifies the DataFrame in place, meaning it doesn't
#return a new DataFrame but rather changes the original DataFrame (data) directly.

# train_dataframe.drop_duplicates(['headline'], keep = 'last', inplace = True)
# test_dataframe.drop_duplicates(['headline'], keep = 'last', inplace = True)
# val_dataframe.drop_duplicates(['headline'], keep = 'last', inplace = True)

# print("Total number of rows after removing duplicates in train = ",train_dataframe.shape[0])
print("Total number of rows after removing duplicates in test  = ",test_dataframe.shape[0])
print("Total number of rows after removing duplicates in val   = ",val_dataframe.shape[0])

# print(train_dataframe.isnull().sum())
print(test_dataframe.isnull().sum())
print(val_dataframe.isnull().sum())

#Blank headlines
# train_dataframe[train_dataframe['headline'] == '']

#Blank headlines
test_dataframe[test_dataframe['headline'] == '']

#Blank headlines
val_dataframe[val_dataframe['headline'] == '']

# train_dataframe = train_dataframe[~(train_dataframe['headline'] == '')]
test_dataframe = test_dataframe[~(test_dataframe['headline'] == '')]
val_dataframe = val_dataframe[~(val_dataframe['headline'] == '')]

# X_train = train_dataframe['headline']
# y_train = train_dataframe['category']

X_test =  test_dataframe['headline']
y_test =  test_dataframe['category']

X_val =  val_dataframe['headline']
y_val =  val_dataframe['category']

# # Convert the list of strings to a DataFrame
# df = pd.DataFrame(list(X_train), columns=['Data'])

# # Specify the CSV file path
# csv_file_path = "NER Spliting Data\\X_train.csv"

# # Write DataFrame to CSV file
# df.to_csv(csv_file_path, index=False)

# # Convert the list of strings to a DataFrame
# df = pd.DataFrame(list(y_train), columns=['Data'])

# # Specify the CSV file path
# csv_file_path = "NER Spliting Data\\y_train.csv"

# # Write DataFrame to CSV file
# df.to_csv(csv_file_path, index=False)

# Convert the list of strings to a DataFrame
df = pd.DataFrame(list(X_test), columns=['Data'])

# Specify the CSV file path
csv_file_path = "NER Spliting Data\\X_test.csv"

# Write DataFrame to CSV file
df.to_csv(csv_file_path, index=False)

# Convert the list of strings to a DataFrame
df = pd.DataFrame(list(y_test), columns=['Data'])

# Specify the CSV file path
csv_file_path = "NER Spliting Data\\y_test.csv"

# Write DataFrame to CSV file
df.to_csv(csv_file_path, index=False)

# Convert the list of strings to a DataFrame
df = pd.DataFrame(list(X_val), columns=['Data'])

# Specify the CSV file path
csv_file_path = "NER Spliting Data\\X_val.csv"

# Write DataFrame to CSV file
df.to_csv(csv_file_path, index=False)

# Convert the list of strings to a DataFrame
df = pd.DataFrame(list(y_val), columns=['Data'])

# Specify the CSV file path
csv_file_path = "NER Spliting Data\\y_val.csv"

# Write DataFrame to CSV file
df.to_csv(csv_file_path, index=False)
