==> 06-Model_Fit_Function_Token.ipynb <==
import pandas as pd
import numpy as np
from sklearn.naive_bayes import MultinomialNB
from sklearn.pipeline import Pipeline
from sklearn.feature_extraction.text import TfidfVectorizer , CountVectorizer
from sklearn.svm import SVC
from sklearn.linear_model import LogisticRegression
from sklearn.feature_selection import SelectPercentile,f_classif
from sklearn.ensemble import RandomForestClassifier , GradientBoostingClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.neighbors import KNeighborsClassifier
from sklearn.preprocessing import StandardScaler
import joblib
import warnings
# Ignore warnings
warnings.filterwarnings("ignore")

X_train = pd.read_csv ('Tokens Spliting Data\\X_train.csv') ['Data']

y_train = pd.read_csv ('Tokens Spliting Data\\y_train.csv') ['Data']

def Model_Fit(model, file_name ):
    global X_train
    global y_train
    
    pipeline = Pipeline([
        ('tfidf', TfidfVectorizer(ngram_range=(1,3))),  # Replace with your desired vectorizer
        ('select_percentile', SelectPercentile(f_classif, percentile=25)),
        ('model', model)
    ])

    pipeline.fit(X_train, y_train)

    joblib.dump( pipeline, file_name )


NB_Model =  MultinomialNB()
Model_Fit(NB_Model , 'Tokens_naive_bayes_pipeline.joblib' )
print ("Naive Bayes is done.")

Log_reg_Model = LogisticRegression()
Model_Fit (Log_reg_Model , 'Tokens_logistic_regression_pipeline.joblib' )
print ("Logisitic regression is done.")

KNN_Model = KNeighborsClassifier(n_neighbors=7)
Model_Fit (KNN_Model , 'NER_knn_pipeline.joblib')
print ("KNeighbors is done.")

SVM_Model = SVC()
Model_Fit ( SVM_Model , 'NER_svm_pipeline.joblib' )
print ("SVM is done.")

RN_Model = RandomForestClassifier(n_estimators=200, random_state=42)
Model_Fit ( RN_Model , 'NER_random_forest_pipeline.joblib' )
print ("Random Forest is done.")

DT_Model = DecisionTreeClassifier()
Model_Fit ( DT_Model , 'NER_DT_pipeline.joblib' )
print ("Decision Tree is done.")

GB_Model = GradientBoostingClassifier()
Model_Fit ( GB_Model , 'NER_gb_pipeline.joblib' )
print ("Gradient Boosting is done.")
